Add recency weighting and team-fallback to the Markov engine, trim verbose comments
The pooled grid->finish transition table now decays older races by 0.85^k
(shared RECENCY_LAMBDA with the Dirichlet model) instead of weighting all
history equally. Drivers with fewer than 10 personal races now fall back
to their team's pooled index instead of getting no personalization at all,
labeled honestly in StrategyReporter's output via a new DriverAdjustment
struct. Also verified position data has no DNF sentinel-value contamination
(clean per-race ranking, no fix needed), and shortened/removed comment
blocks over 3 lines across the engine, service, and test files.

+++ markov_engine.h
@@ -3,10 +3,10 @@
 
 class MarkovEngine {
 private:
-    const std::map<int, std::map<int, int>>& counts;
+    const std::map<int, std::map<int, double>>& counts;
 
 public:
-    explicit MarkovEngine(const std::map<int, std::map<int, int>>& transition_counts);
+    explicit MarkovEngine(const std::map<int, std::map<int, double>>& transition_counts);
 
     std::map<int, double> predict_finish_distribution(int grid_position) const;
     int most_likely_finish(int grid_position) const;
